Remove event retargeting work-around.
This event retarget was originally added in order to work around a
problem with react event propagation across the shadow root boundary.
This appears to be no longer needed.  It is currently resulting in
double dispatching to the on click handler.

While not yet confirmed, it is possible that the underlying issue was
fixed with https://github.com/facebook/react/pull/11927 which was
released in react-dom v16.5.0.